Rename cbf_computation to cbf.
